# docs/swagger.yaml
definitions:
...
  go-file-server_internal_model.CloudConfig:
    properties:
      config_name:
        type: string
      config_type:
        type: string
      config_unit:
        type: string
      config_value:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      is_enabled:
        type: boolean
    type: object
...
  internal_controller.UpdateConfigRequest:
    properties:
      config_value:
        type: string
      is_deleted:
        type: boolean
      is_enabled:
        type: boolean
    type: object
...
host: localhost:3333
schemes:
  - http
info:
  contact:
    name: GoNet Drive Support
    url: https://github.com
  description: GoNet Drive is a self-hosted cloud file storage and sharing service.
  termsOfService: https://example.com/terms
  title: GoNet Drive API
  version: "1.0"
paths:
...
  /api/user/admin/config/logo:
    put:
      consumes:
      - multipart/form-data
      description: Upload a new PNG logo file. Requires admin role.
      parameters:
      - description: PNG logo file
        in: formData
        name: logo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Update Logo
      tags:
      - Admin
...
  /api/user/config:
    get:
      description: Get all cloud configuration entries.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/go-file-server_internal_model.CloudConfig'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: List Configs
      tags:
      - Config
  /api/user/config/{id}:
    put:
      consumes:
      - application/json
      description: Update a cloud configuration entry (value, enabled status, or soft-delete).
      parameters:
      - description: Config ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_controller.UpdateConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Update Config
      tags:
      - Config
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: JWT access token stored as a cookie.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"